requirements.txt
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-limiter==0.1.6
httpx==0.27.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langchain-together==0.1.4
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==4.3.1
pytest==8.3.1
python-dotenv==1.0.1
python-multipart==0.0.9
sentence-transformers==3.0.1
uvicorn==0.30.3